+++ src/setup/setup.h
@@ -93,15 +93,15 @@ extern "C" {
 C_KZG_RET load_trusted_setup(
     KZGSettings *out,
     const uint8_t *g1_monomial_bytes,
-    size_t num_g1_monomial_bytes,
+    uint64_t num_g1_monomial_bytes,
     const uint8_t *g1_lagrange_bytes,
-    size_t num_g1_lagrange_bytes,
+    uint64_t num_g1_lagrange_bytes,
     const uint8_t *g2_monomial_bytes,
-    size_t num_g2_monomial_bytes,
-    size_t precompute
+    uint64_t num_g2_monomial_bytes,
+    uint64_t precompute
 );
 
-C_KZG_RET load_trusted_setup_file(KZGSettings *out, FILE *in, size_t precompute);
+C_KZG_RET load_trusted_setup_file(KZGSettings *out, FILE *in, uint64_t precompute);
 
 void free_trusted_setup(KZGSettings *s);
